docs/conf.py
from docs_conf.conf import *

#branch configuration

branch = 'latest'

linkcheck_ignore = [
    'http://localhost.*',
    'http://127.0.0.1.*',
    'https://gerrit.o-ran-sc.org.*',
]

extensions = ['sphinxcontrib.redoc', 'sphinx.ext.intersphinx',]
